Apply transformations from dhis2 to domain
