+++ src/logic/sync/events.ts
@@ -26,7 +26,7 @@ import {
 } from "../../utils/synchronization";
 import { AggregatedSync } from "./aggregated";
 import { GenericSync, SyncronizationPayload } from "./generic";
-import { mapPackageToD2Version } from "../../data/synchronization/mappers/D2VersionPackageMapper";
+import { mapPackageToD2 } from "../../data/synchronization/mappers/PackageMapper";
 import { eventsTransformationsToDhis2 } from "../../data/synchronization/mappers/PackageTransformations";
 
 export class EventsSync extends GenericSync {
@@ -92,7 +92,7 @@ export class EventsSync extends GenericSync {
             throw new Error("Necessary api version of receiver instance to apply transformations to package is undefined")
         }
 
-        const versionedPayloadPackage = mapPackageToD2Version(instanceEntity.apiVersion, payload, eventsTransformationsToDhis2);
+        const versionedPayloadPackage = mapPackageToD2(instanceEntity.apiVersion, payload, eventsTransformationsToDhis2);
         console.debug("Events package", { events, payload, versionedPayloadPackage });
 
         const response = await postEventsData(instance, payload, dataParams);
